pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "flitter-kinect"
version = "1.0"
description = "A Kinect v2 window render plugin for flitter-lang"
authors = [
  {name = "Jonathan Hogg", email = "me@jonathanhogg.com"},
]
maintainers = [
  {name = "Jonathan Hogg", email = "me@jonathanhogg.com"},
]
readme = "README.md"
license = {file="LICENSE"}
keywords = ["language", "opengl", "live-coding", "live-visuals"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: BSD License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Environment :: Console",
  "Environment :: GPU",
  "Environment :: MacOS X",
  "Environment :: X11 Applications",
  "Topic :: Artistic Software",
  "Topic :: Multimedia :: Graphics",
  "Topic :: Multimedia :: Graphics :: 3D Rendering",
]
requires-python = ">= 3.10"
dependencies = [
    "flitter-lang",
    "freenect2",
    "loguru",
    "Mako",
    "numpy",
    "pyusb",
    "trimesh",
]

[project.urls]
Documentation = "https://flitter.readthedocs.io/"
Repository = "https://github.com/jonathanhogg/flitter-kinect.git"
Issues = "https://github.com/jonathanhogg/flitter-kinect/issues"

[project.entry-points."flitter.render.window"]
kinect = "flitter_kinect.shader:Kinect"

[project.entry-points."flitter.render.window.models"]
kinect = "flitter_kinect.mesh:Kinect"

[tool.setuptools.packages.find]
where = ["src"]

[tool.cython-lint]
max-line-length = 160
